Use SQLite & re-use GTFS IDs

+++ src/main/java/org/mtransit/parser/gtfs/data/GCalendarDate.kt
@@ -68,6 +68,6 @@ data class GCalendarDate(
         fun getNewUID(
             date: Int,
             serviceIdInt: Int
-        ) = "${date}0${serviceIdInt}"
+        ) = "${date}0${serviceIdInt}".toLong()
     }
 }
